[FIX] mis_builder: Be tolerant in tests with several installed companies
If you are running tests in a integrated environment, you can have more
than one company with the same account 200x, so we need to filter on the
current active company.

We may have as well more companies as authorized ones, so let's assert
the proper thing in `test_query_company_ids`.