[IMP] Rework the analytic filtering mechanism.
Before it was quite ad-hoc with dedicated filters with
analytic account and analytic tags.

Now it is based on a generic analytic_domain filter,
which can be set at the report level or at the column level,
as well as with a search bar in the preview widget.

Additionally, we remove the possibility to override
the analytic filter with the context,
as it was a potential security issue.
Now the context can contain a mis_analytic_domain
key (replacing the mis_report_filters key),  but
this domain is always ANDed with the domains
configured on the report and columns.

The report widget also gets some code quality improvements.