[FIX][18.0] mis_builder: Error in multicompany report
Description of the issue/feature this PR addresses:

Current behavior before PR:

When we have two companies selected, and we go to Invoicing > Reporting > MIS reports, select a report template, check ‘Multiple companies’ and click on preview, we get an error message.

Desired behavior after PR is merged:

The error no longer appears and the preview is displayed correctly.

ADD:
When we have multiple companies but only one is selected in the context,
the report should not display the company name. This fix ensures the company
name is only shown when more than one company is active.

@moduon MT-11000